Moves PipelineBuffer to pipeline namespace
