+++ src/pipeline/DecodeUnit.hh
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../Architecture.hh"
-#include "../PipelineBuffer.hh"
+#include "PipelineBuffer.hh"
 
 namespace simeng {
 namespace pipeline {
