+++ src/pipeline/WritebackUnit.hh
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../Instruction.hh"
-#include "../PipelineBuffer.hh"
+#include "PipelineBuffer.hh"
 
 namespace simeng {
 namespace pipeline {
